careplan-mvp/requirements.txt
django==5.1
djangorestframework==3.15.2
psycopg2-binary==2.9.9
anthropic==0.40.0
django-cors-headers==4.4.0
